src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/constructor-calls-created-synchronously-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: document.registerElement is not a function
Tests that custom element constructors run the createdCallback synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
